requirements.txt
ultralytics
opencv-python
streamlit
Pillow
numpy